+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Migrate [app-tree-utils](https://github.com/badoo/app-tree-utils) into this repository.
  - Do not allow setting `Node.integrationPoint` on non-root nodes.
  - #20 Fix integration point attached twice crash when using live literals 
+ - Fix transition interruptions bug
 
 ## 1.0-alpha01